+++ types/events.d.ts
@@ -35,7 +35,7 @@ export type ClipboardEvent = {
 	 */
 	maxCharCount: boolean;
 	/**
-	 * - `"SE"`|`"MS"`|`""` - source
+	 * - `"SE"`|`"MS"`|`"GOOGLE"`|`""` - source
 	 */
 	from: string;
 };
@@ -289,7 +289,7 @@ export type EventHandlers = {
  * @property {Event} event - event object
  * @property {string} data - drop data
  * @property {boolean} maxCharCount - is max char count
- * @property {string} from - `"SE"`|`"MS"`|`""` - source
+ * @property {string} from - `"SE"`|`"MS"`|`"GOOGLE"`|`""` - source
  */
 /**
  * @typedef {Object} FileManagementInfo